requirements.txt
speedtest-cli==2.1.3
colorama==0.4.4
matplotlib==3.4.3
speedtest-cli